estrategico/etl_xp_hub/requirements.txt
# Dependências necessárias
pandas>=1.5.0
openpyxl>=3.0.0
sqlalchemy>=1.4.0
pyodbc>=4.0.0
boto3>=1.26.0
python-dotenv>=0.19.0
apache-airflow>=2.5.0
